Update StockController.cs
Added Http Put method:
async Task<ActionResult> PutBook(BookProduct). !ATTENTION! if new genre/genres was entered then the all existing genres would be replaced with new entered list. Required: Id, isAvailable another parameters could be not entered, parameters that was not entered would be not changed. If Id was not specified or incorrect - returns BadRequest(), if Id is not found - returns NotFound() object result. If isAvailable not specified - sets to false.